rpk: move cmd/iotune to its own package
